Typo (`LICENCE` -> `LICENSE`), user-dismissible UI notice, missing i18n.
The `LICENCE` file is renamed to `LICENSE` because the latter is the
correct English spelling of the word.

The admin notice upon plugin activation now supports a user-dismissible
button, available from WordPress 4.2 and higher and backwards-compatible
with previous versions of WordPress. For details, see:

https://make.wordpress.org/core/2015/04/23/spinners-and-dismissible-admin-notices-in-4-2/

Finally, the "Install now" string was not translatable/localizeable
because it was hardcoded into the admin UI.

The `languages/messages.pot` file is also updated, because many such
strings were missing from it.