Fixes #22: Only show the "no database installed" admin notice to admins.
This commit also adds a `checkRequirements()` method to the
`WP_UnitTestCase_GeoIP_Detect` class in order to ensure that Travis-CI
does not attempt to fetch WordPress Core Trac tickets when building.
This is done in order to make our own use of the `@ticket` annotation in
PHPUnit to refer to issues on GitHub.com.

This commit includes test cases for the bugfix.