Fix unit tests for older version of WordPress.
Older versions of WordPress include the `WP_UnitTest_Factory` but don't
automatically assign an instance to the `WP_UnitTestCase::$factory`
member variable. We have to do that ourselves. Also, older versions of
the `WP_UnitTest_Factory` class don't do a good job of recreating user
logins or email addresses, so we randomize those ourselves, as well.